op_alloy_rpc_types_engine::superchain
Enum ProtocolVersion

#[non_exhaustive]
pub enum ProtocolVersion {
    V0(ProtocolVersionFormatV0),
}

Formatted Superchain Protocol Version.

The Protocol Version documents the progression of the total set of canonical OP-Stack specifications. Components of the OP-Stack implement the subset of their respective protocol component domain, up to a given Protocol Version of the OP-Stack.

The Protocol Version is NOT a hardfork identifier, but rather indicates software-support for a well-defined set of features introduced in past and future hardforks, not the activation of said hardforks.

The Protocol Version is Semver-compatible. It is encoded as a single 32 bytes long protocol version. The version must be encoded as 32 bytes of DATA in JSON RPC usage.

See also: https://specs.optimism.io/protocol/superchain-upgrades.html#protocol-version

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

V0(ProtocolVersionFormatV0)

Version-type 0.

Implementations

impl ProtocolVersion

pub fn encode(&self) -> B256

Version-type 0 byte encoding:

<protocol version> ::= <version-type><typed-payload>
<version-type> ::= <uint8>
<typed-payload> ::= <31 bytes>

pub fn decode(value: B256) -> Result<Self, ProtocolVersionError>

Version-type 0 byte decoding:

<protocol version> ::= <version-type><typed-payload>
<version-type> ::= <uint8>
<typed-payload> ::= <31 bytes>

pub const fn inner(&self) -> ProtocolVersionFormatV0

Returns the inner value of the ProtocolVersion enum

pub const fn as_v0(&self) -> Option<ProtocolVersionFormatV0>

Returns the inner value of the ProtocolVersion enum if it is V0, otherwise None

pub const fn build(&self) -> B64

Differentiates forks and custom-builds of standard protocol

pub const fn major(&self) -> u32

Incompatible API changes

pub const fn minor(&self) -> u32

Identifies additional functionality in backwards compatible manner

pub const fn patch(&self) -> u32

Identifies backward-compatible bug-fixes

pub const fn pre_release(&self) -> u32

Identifies unstable versions that may not satisfy the above

pub fn display(&self) -> String

Returns a human-readable string representation of the ProtocolVersion