op_alloy_rpc_types_engine
Trait OptimismPayloadUtils

pub trait OptimismPayloadUtils {
    // Required methods
    fn to_l2_block_info(
        &self,
        genesis: &ChainGenesis,
    ) -> Result<L2BlockInfo, ToL2BlockRefError>;
    fn to_system_config(
        &self,
        genesis: &ChainGenesis,
    ) -> Result<SystemConfig, ToSystemConfigError>;
}

Defines conversion utility methods for Optimism-specific payloads.

Required Methods

fn to_l2_block_info( &self, genesis: &ChainGenesis, ) -> Result<L2BlockInfo, ToL2BlockRefError>

Converts the payload into an L2BlockInfo.

This method creates the L2BlockInfo using the inner payload fetched with the AsInnerPayload::as_v1_payload method.

If the payload is the genesis block, it will return the genesis block info. Otherwise, it will return the L2 block info from the first deposit transaction.

fn to_system_config( &self, genesis: &ChainGenesis, ) -> Result<SystemConfig, ToSystemConfigError>

Converts the payload into a SystemConfig.

This method creates the SystemConfig using the inner payload fetched with the AsInnerPayload::as_v1_payload method.

If the payload is the genesis block, it will return the system config from the genesis block. Otherwise, it will return the system config from the first deposit transaction.