Crate op_alloy_rpc_types_engine

op-alloy-rpc-types-engine

Optimism RPC types for the engine namespace“

Modules

attributes 🔒

Optimism-specific payload attributes.

errors 🔒

Error types for conversions between Optimism Execution Payload Envelope types and external types.

payload_v3 🔒

Optimism execution payload envelope V3.

payload_v4 🔒

Optimism execution payload envelope V3.

superchain 🔒

Structs

ChainGenesis

Chain genesis information.

ExecutionPayloadV1

This structure maps on the ExecutionPayload structure of the beacon chain spec.

L2BlockInfo

L2 Block Header Info

OptimismAttributesWithParent

Optimism Payload Attributes with parent block reference.

OptimismExecutionPayloadEnvelopeV3

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V3.

OptimismExecutionPayloadEnvelopeV4

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V4.

OptimismPayloadAttributes

Optimism Payload Attributes

ProtocolVersionFormatV0

The Protocol Version V0 format. Encoded as 31 bytes with the following structure:

RollupConfig

The Rollup configuration.

SuperchainSignal

Superchain Signal information.

SystemConfig

System configuration.

Enums

ProtocolVersion

Formatted Superchain Protocol Version.

ProtocolVersionError

An error that can occur when encoding or decoding a ProtocolVersion.

ToL2BlockRefError

An error that can occur when converting an crate::OptimismExecutionPayloadEnvelopeV4 to an op_alloy_protocol::L2BlockInfo.

ToSystemConfigError

An error that can occur when converting an crate::OptimismExecutionPayloadEnvelopeV4 to a op_alloy_genesis::SystemConfig.

Traits

AsInnerPayload

Converts a payload into its most inner ExecutionPayloadV1.

OptimismPayloadUtils

Defines conversion utility methods for Optimism-specific payloads.